Add a session feature to disable reqwest for use as generic type base
