+++ Cargo.toml
@@ -11,27 +11,27 @@ readme = "README.md"
 aho-corasick = "1.1"
 base64 = "0.22"
 chrono = "0.4"
-enum-map = "2.7.3"
+enum-map = "2.7"
 fastrand = "2.2"
 libaes = "0.7"
 log = "0.4"
-num = "0.4"
-num-bigint = { version = "0.4" }
+num-bigint = "0.4"
 num-derive = "0.4"
 num-traits = "0.2"
 once_cell = "1.20"
 ouroboros = "0.18"
-reqwest = { version = "0.12", features = ["gzip", "deflate", "brotli"] }
+reqwest = { version = "0.12", features = ["gzip", "deflate"], optional = true }
 serde = { version = "1.0", features = ["derive"], optional = true }
 serde_json = { version = "1.0", optional = true }
 sha1 = "0.10"
 strum = { version = "0.26", features = ["strum_macros", "derive"] }
-tokio = "1.41"
-url = "2.5"
+tokio = { version = "1.41", optional = true }
+url = { version = "2.5", optional = true }
 
 [features]
 default = ["sso", "serde"]
-sso = ["dep:serde", "dep:serde_json"]
+session = ["dep:reqwest", "dep:tokio", "dep:url"]
+sso = ["session", "dep:serde", "dep:serde_json"]
 serde = ["dep:serde", "num-bigint/serde", "chrono/serde", "enum-map/serde"]
 
 [dev-dependencies]
