+++ examples/expedition.rs
@@ -4,7 +4,7 @@ use chrono::{DateTime, Local};
 use sf_api::{
     command::{Command, ExpeditionSetting, TimeSkip},
     gamestate::tavern::{AvailableTasks, ExpeditionStage},
-    SimpleSession,
+    session::SimpleSession,
 };
 use tokio::time::sleep;
 
@@ -149,7 +149,7 @@ pub async fn login_with_env() -> SimpleSession {
     let username = std::env::var("USERNAME").unwrap();
     let password = std::env::var("PASSWORD").unwrap();
     let server = std::env::var("SERVER").unwrap();
-    sf_api::SimpleSession::login(&username, &password, &server)
+    SimpleSession::login(&username, &password, &server)
         .await
         .unwrap()
 }
