+++ src/gamestate/mod.rs
@@ -28,7 +28,7 @@ use crate::{
         items::*, rewards::*, social::*, tavern::*, unlockables::*,
     },
     misc::*,
-    session::*,
+    response::Response,
 };
 
 #[derive(Debug, Clone, Default)]
