+++ src/response.rs
@@ -1,0 +1,312 @@
+use std::{collections::HashMap, fmt::Debug, str::FromStr};
+
+use chrono::NaiveDateTime;
+use log::{error, trace, warn};
+
+use crate::error::SFError;
+
+#[ouroboros::self_referencing]
+/// A bunch of new information about the state of the server and/or the
+/// player
+///
+/// NOTE: This has a weird syntax to access, because we do not want to create
+/// 10000 strings on each request and instead just store the raw response body
+/// and references into it. This is faster & uses less memory, but because of
+/// rusts borrow checker requires some weird syntax here.
+// Technically we could do this safely with an iterator, that parses on demand,
+// but send_command() needs to access specifix response keys to keep the session
+// running, which means a hashmap needs to be constructed no matter what
+pub struct Response {
+    body: String,
+    #[borrows(body)]
+    #[covariant]
+    resp: HashMap<&'this str, ResponseVal<'this>>,
+    /// We store this to make sure the time calculations are still correct, if
+    /// this response is held any amount of time before being used to update
+    /// character state
+    received_at: NaiveDateTime,
+}
+
+impl Clone for Response {
+    // This is not a good clone..
+    #[allow(clippy::expect_used)]
+    fn clone(&self) -> Self {
+        Self::parse(self.raw_response().to_string(), self.received_at())
+            .expect("Invalid response cloned")
+    }
+}
+
+impl Debug for Response {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        f.debug_map()
+            .entries(self.values().iter().map(|a| (a.0, a.1.as_str())))
+            .finish()
+    }
+}
+
+#[cfg(feature = "serde")]
+impl serde::Serialize for Response {
+    fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
+    where
+        S: serde::Serializer,
+    {
+        use serde::ser::SerializeStruct;
+        let mut s = serializer.serialize_struct("Response", 2)?;
+        s.serialize_field("body", self.borrow_body())?;
+        s.serialize_field("received_at", &self.received_at())?;
+        s.end()
+    }
+}
+
+#[cfg(feature = "serde")]
+impl<'de> serde::Deserialize<'de> for Response {
+    fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
+    where
+        D: serde::Deserializer<'de>,
+    {
+        struct AVisitor;
+
+        impl<'de> serde::de::Visitor<'de> for AVisitor {
+            type Value = Response;
+
+            fn expecting(
+                &self,
+                formatter: &mut std::fmt::Formatter,
+            ) -> std::fmt::Result {
+                formatter.write_str(
+                    "struct Response with fields body and received_at",
+                )
+            }
+
+            fn visit_map<V>(self, mut map: V) -> Result<Self::Value, V::Error>
+            where
+                V: serde::de::MapAccess<'de>,
+            {
+                let mut body = None;
+                let mut received_at = None;
+
+                while let Some(key) = map.next_key()? {
+                    match key {
+                        "body" => {
+                            body = Some(map.next_value()?);
+                        }
+                        "received_at" => {
+                            received_at = Some(map.next_value()?);
+                        }
+                        _ => {
+                            // Ignore unknown fields
+                            map.next_value::<serde::de::IgnoredAny>()?;
+                        }
+                    }
+                }
+
+                let body: String =
+                    body.ok_or_else(|| serde::de::Error::missing_field("q"))?;
+                let received_at: NaiveDateTime = received_at
+                    .ok_or_else(|| serde::de::Error::missing_field("j"))?;
+
+                Response::parse(body, received_at).map_err(|_| {
+                    serde::de::Error::custom("invalid resopnse body")
+                })
+            }
+        }
+
+        deserializer.deserialize_struct(
+            "Response",
+            &["body", "received_at"],
+            AVisitor,
+        )
+    }
+}
+
+impl Response {
+    /// Returns a reference to the hashmap, that contains mappings of response
+    /// keys to values
+    #[must_use]
+    pub fn values(&self) -> &HashMap<&str, ResponseVal<'_>> {
+        self.borrow_resp()
+    }
+
+    /// Returns the raw response from the server. This should only ever be
+    /// necessary for debugging, caching, or in case there is ever a new
+    /// response format in a response, that is not yet supported. You can of
+    /// course also use this to look at how horrible the S&F encoding is..
+    #[must_use]
+    pub fn raw_response(&self) -> &str {
+        self.borrow_body()
+    }
+
+    /// Returns the time, at which the response was received
+    #[must_use]
+    pub fn received_at(&self) -> NaiveDateTime {
+        self.with_received_at(|a| *a)
+    }
+
+    /// Parses a response body from the server into a useable format
+    /// You might want to use this, if you are analyzing reponses from the
+    /// browsers network tab. If you are trying to store/read responses to/from
+    /// disk to cache them, or otherwise, you should use the sso feature to
+    /// serialize/deserialize them instead
+    ///
+    /// # Errors
+    /// - `ServerError`: If the server responsed with an error
+    /// - `ParsingError`: If the reponse does not follow the standard S&F server
+    ///   response schema
+    pub fn parse(
+        og_body: String,
+        received_at: NaiveDateTime,
+    ) -> Result<Response, SFError> {
+        // We can not return from the closure below, so we have to do this work
+        // twice (sadly)
+
+        // NOTE: I think the trims might actually be completely unnecessary.
+        // Pretty sure I mixed them up with command encoding, which is actually
+        // '|' padded
+
+        let body = og_body
+            .trim_end_matches('|')
+            .trim_start_matches(|a: char| !a.is_alphabetic());
+        trace!("Received raw response: {body}");
+
+        if !body.contains(':')
+            && !body.starts_with("success")
+            && !body.starts_with("Success")
+        {
+            return Err(SFError::ParsingError(
+                "unexpected server response",
+                body.to_string(),
+            ));
+        }
+
+        if body.starts_with("error") || body.starts_with("Error") {
+            let raw_error = body.split_once(':').unwrap_or_default().1;
+
+            let error_msg = match raw_error {
+                "adventure index must be 1-3" => "quest index must be 0-2",
+                x => x,
+            };
+
+            return Err(SFError::ServerError(error_msg.to_string()));
+        }
+
+        let resp = ResponseBuilder {
+            body: og_body,
+            resp_builder: |body: &String| {
+                let mut res = HashMap::new();
+                for part in body
+                    .trim_start_matches(|a: char| !a.is_alphabetic())
+                    .trim_end_matches('|')
+                    .split('&')
+                    .filter(|a| !a.is_empty())
+                {
+                    let Some((full_key, value)) = part.split_once(':') else {
+                        warn!("weird k/v in resp: {part}");
+                        continue;
+                    };
+
+                    let (key, sub_key) = match full_key.split_once('.') {
+                        Some(x) => {
+                            // full_key == key.subkey
+                            x
+                        }
+                        None => {
+                            if let Some((k, sk)) = full_key.split_once('(') {
+                                // full_key == key(4)
+                                (k, sk.trim_matches(')'))
+                            } else {
+                                // full_key == key
+                                (full_key, "")
+                            }
+                        }
+                    };
+                    if key.is_empty() {
+                        continue;
+                    }
+
+                    res.insert(key, ResponseVal { value, sub_key });
+                }
+                res
+            },
+            received_at,
+        }
+        .build();
+
+        Ok(resp)
+    }
+}
+
+#[derive(Debug, Clone, Copy)]
+/// This is the raw &str, that the server send as a value to some key. This
+/// often requires extra conversions/parsing to use practically, so we associate
+/// the most common parsing functions as methods to this data.
+pub struct ResponseVal<'a> {
+    value: &'a str,
+    sub_key: &'a str,
+}
+
+impl<'a> ResponseVal<'a> {
+    /// Converts the response value into the required type
+    ///
+    /// # Errors
+    /// If the reponse value can not be parsed into the output
+    /// value, a `ParsingError` will be returned
+    pub fn into<T: FromStr>(self, name: &'static str) -> Result<T, SFError> {
+        self.value.trim().parse().map_err(|_| {
+            error!("Could not convert {name} into target type: {self}");
+            SFError::ParsingError(name, self.value.to_string())
+        })
+    }
+
+    /// Converts the repsponse into a list, by splitting the raw value by '/'
+    /// and converting each value into the required type. If any conversion
+    /// fails, an error is returned
+    ///
+    /// # Errors
+    /// If any of the values in the string can not be parsed into the output
+    /// value, the `ParsingError` for that value will be returned
+    pub fn into_list<T: FromStr>(
+        self,
+        name: &'static str,
+    ) -> Result<Vec<T>, SFError> {
+        let x = &self.value;
+        if x.is_empty() {
+            return Ok(Vec::new());
+        }
+        // Trimming ` ` & `\n` is not required. Might remove this later
+        x.trim_matches(|a| ['/', ' ', '\n'].contains(&a))
+            .split('/')
+            .map(|c| {
+                c.trim().parse::<T>().map_err(|_| {
+                    error!(
+                        "Could not convert {name} into list because of {c}: \
+                         {self}"
+                    );
+                    SFError::ParsingError(name, format!("{c:?}"))
+                })
+            })
+            .collect()
+    }
+
+    /// The way keys are parsed will trim some info from the string. The key for
+    /// the player save `ownplayersave` is actually `ownplayersave.playerSave`.
+    /// As this `.playerSave` is not relevant here and not in most cases, I
+    /// decided to trim that off. More common, this is also just `s`, `r`, or a
+    /// size hint like `(10)`. In some cases though, this information can be
+    /// helpful for parsing. Thus, you can access it here
+    #[must_use]
+    pub fn sub_key(&self) -> &str {
+        self.sub_key
+    }
+
+    /// Returns the raw reference to the internal &str, that the server send
+    #[must_use]
+    pub fn as_str(&self) -> &str {
+        self.value
+    }
+}
+
+impl<'a> std::fmt::Display for ResponseVal<'a> {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        f.write_str(self.value)
+    }
+}
