+++ src/session.rs
@@ -1,13 +1,10 @@
 use std::{
     borrow::Borrow,
-    collections::HashMap,
     fmt::Debug,
-    str::FromStr,
     sync::{atomic::AtomicU32, Arc},
+    time::Duration,
 };
 
-use base64::Engine;
-use chrono::NaiveDateTime;
 use log::{error, trace, warn};
 use reqwest::{header::*, Client};
 use url::Url;
@@ -15,14 +12,16 @@ use url::Url;
 use crate::{
     command::Command,
     error::SFError,
-    gamestate::character::{Class, Gender, Race},
-    misc::{sha1_hash, HASH_CONST},
+    gamestate::{
+        character::{Class, Gender, Race},
+        GameState,
+    },
+    misc::{
+        encrypt_server_request, sha1_hash, DEFAULT_CRYPTO_ID,
+        DEFAULT_CRYPTO_KEY, DEFAULT_SESSION_ID, HASH_CONST,
+    },
 };
-
-pub(crate) const DEFAULT_CRYPTO_KEY: &str = "[_/$VV&*Qg&)r?~g";
-pub(crate) const DEFAULT_CRYPTO_ID: &str = "0-00000000000000";
-pub(crate) const DEFAULT_SESSION_ID: &str = "00000000000000000000000000000000";
-const CRYPTO_IV: &str = "jXT#/vz]3]5X7Jl\\";
+pub use crate::{misc::decrypt_url, response::*};
 
 #[derive(Debug, Clone)]
 /// The session, that manages the server communication for a character
@@ -416,312 +415,6 @@ impl Session {
     }
 }
 
-#[ouroboros::self_referencing]
-/// A bunch of new information about the state of the server and/or the
-/// player
-///
-/// NOTE: This has a weird syntax to access, because we do not want to create
-/// 10000 strings on each request and instead just store the raw response body
-/// and references into it. This is faster & uses less memory, but because of
-/// rusts borrow checker requires some weird syntax here.
-// Technically we could do this safely with an iterator, that parses on demand,
-// but send_command() needs to access specifix response keys to keep the session
-// running, which means a hashmap needs to be constructed no matter what
-pub struct Response {
-    body: String,
-    #[borrows(body)]
-    #[covariant]
-    resp: HashMap<&'this str, ResponseVal<'this>>,
-    /// We store this to make sure the time calculations are still correct, if
-    /// this response is held any amount of time before being used to update
-    /// character state
-    received_at: NaiveDateTime,
-}
-
-impl Clone for Response {
-    // This is not a good clone..
-    #[allow(clippy::expect_used)]
-    fn clone(&self) -> Self {
-        Self::parse(self.raw_response().to_string(), self.received_at())
-            .expect("Invalid response cloned")
-    }
-}
-
-impl Debug for Response {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        f.debug_map()
-            .entries(self.values().iter().map(|a| (a.0, a.1.as_str())))
-            .finish()
-    }
-}
-
-#[cfg(feature = "serde")]
-impl serde::Serialize for Response {
-    fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
-    where
-        S: serde::Serializer,
-    {
-        use serde::ser::SerializeStruct;
-        let mut s = serializer.serialize_struct("Response", 2)?;
-        s.serialize_field("body", self.borrow_body())?;
-        s.serialize_field("received_at", &self.received_at())?;
-        s.end()
-    }
-}
-
-#[cfg(feature = "serde")]
-impl<'de> serde::Deserialize<'de> for Response {
-    fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
-    where
-        D: serde::Deserializer<'de>,
-    {
-        struct AVisitor;
-
-        impl<'de> serde::de::Visitor<'de> for AVisitor {
-            type Value = Response;
-
-            fn expecting(
-                &self,
-                formatter: &mut std::fmt::Formatter,
-            ) -> std::fmt::Result {
-                formatter.write_str(
-                    "struct Response with fields body and received_at",
-                )
-            }
-
-            fn visit_map<V>(self, mut map: V) -> Result<Self::Value, V::Error>
-            where
-                V: serde::de::MapAccess<'de>,
-            {
-                let mut body = None;
-                let mut received_at = None;
-
-                while let Some(key) = map.next_key()? {
-                    match key {
-                        "body" => {
-                            body = Some(map.next_value()?);
-                        }
-                        "received_at" => {
-                            received_at = Some(map.next_value()?);
-                        }
-                        _ => {
-                            // Ignore unknown fields
-                            map.next_value::<serde::de::IgnoredAny>()?;
-                        }
-                    }
-                }
-
-                let body: String =
-                    body.ok_or_else(|| serde::de::Error::missing_field("q"))?;
-                let received_at: NaiveDateTime = received_at
-                    .ok_or_else(|| serde::de::Error::missing_field("j"))?;
-
-                Response::parse(body, received_at).map_err(|_| {
-                    serde::de::Error::custom("invalid resopnse body")
-                })
-            }
-        }
-
-        deserializer.deserialize_struct(
-            "Response",
-            &["body", "received_at"],
-            AVisitor,
-        )
-    }
-}
-
-impl Response {
-    /// Returns a reference to the hashmap, that contains mappings of response
-    /// keys to values
-    #[must_use]
-    pub fn values(&self) -> &HashMap<&str, ResponseVal<'_>> {
-        self.borrow_resp()
-    }
-
-    /// Returns the raw response from the server. This should only ever be
-    /// necessary for debugging, caching, or in case there is ever a new
-    /// response format in a response, that is not yet supported. You can of
-    /// course also use this to look at how horrible the S&F encoding is..
-    #[must_use]
-    pub fn raw_response(&self) -> &str {
-        self.borrow_body()
-    }
-
-    /// Returns the time, at which the response was received
-    #[must_use]
-    pub fn received_at(&self) -> NaiveDateTime {
-        self.with_received_at(|a| *a)
-    }
-
-    /// Parses a response body from the server into a useable format
-    /// You might want to use this, if you are analyzing reponses from the
-    /// browsers network tab. If you are trying to store/read responses to/from
-    /// disk to cache them, or otherwise, you should use the sso feature to
-    /// serialize/deserialize them instead
-    ///
-    /// # Errors
-    /// - `ServerError`: If the server responsed with an error
-    /// - `ParsingError`: If the reponse does not follow the standard S&F server
-    ///   response schema
-    pub fn parse(
-        og_body: String,
-        received_at: NaiveDateTime,
-    ) -> Result<Response, SFError> {
-        // We can not return from the closure below, so we have to do this work
-        // twice (sadly)
-
-        // NOTE: I think the trims might actually be completely unnecessary.
-        // Pretty sure I mixed them up with command encoding, which is actually
-        // '|' padded
-
-        let body = og_body
-            .trim_end_matches('|')
-            .trim_start_matches(|a: char| !a.is_alphabetic());
-        trace!("Received raw response: {body}");
-
-        if !body.contains(':')
-            && !body.starts_with("success")
-            && !body.starts_with("Success")
-        {
-            return Err(SFError::ParsingError(
-                "unexpected server response",
-                body.to_string(),
-            ));
-        }
-
-        if body.starts_with("error") || body.starts_with("Error") {
-            let raw_error = body.split_once(':').unwrap_or_default().1;
-
-            let error_msg = match raw_error {
-                "adventure index must be 1-3" => "quest index must be 0-2",
-                x => x,
-            };
-
-            return Err(SFError::ServerError(error_msg.to_string()));
-        }
-
-        let resp = ResponseBuilder {
-            body: og_body,
-            resp_builder: |body: &String| {
-                let mut res = HashMap::new();
-                for part in body
-                    .trim_start_matches(|a: char| !a.is_alphabetic())
-                    .trim_end_matches('|')
-                    .split('&')
-                    .filter(|a| !a.is_empty())
-                {
-                    let Some((full_key, value)) = part.split_once(':') else {
-                        warn!("weird k/v in resp: {part}");
-                        continue;
-                    };
-
-                    let (key, sub_key) = match full_key.split_once('.') {
-                        Some(x) => {
-                            // full_key == key.subkey
-                            x
-                        }
-                        None => {
-                            if let Some((k, sk)) = full_key.split_once('(') {
-                                // full_key == key(4)
-                                (k, sk.trim_matches(')'))
-                            } else {
-                                // full_key == key
-                                (full_key, "")
-                            }
-                        }
-                    };
-                    if key.is_empty() {
-                        continue;
-                    }
-
-                    res.insert(key, ResponseVal { value, sub_key });
-                }
-                res
-            },
-            received_at,
-        }
-        .build();
-
-        Ok(resp)
-    }
-}
-
-#[derive(Debug, Clone, Copy)]
-/// This is the raw &str, that the server send as a value to some key. This
-/// often requires extra conversions/parsing to use practically, so we associate
-/// the most common parsing functions as methods to this data.
-pub struct ResponseVal<'a> {
-    value: &'a str,
-    sub_key: &'a str,
-}
-
-impl<'a> ResponseVal<'a> {
-    /// Converts the response value into the required type
-    ///
-    /// # Errors
-    /// If the reponse value can not be parsed into the output
-    /// value, a `ParsingError` will be returned
-    pub fn into<T: FromStr>(self, name: &'static str) -> Result<T, SFError> {
-        self.value.trim().parse().map_err(|_| {
-            error!("Could not convert {name} into target type: {self}");
-            SFError::ParsingError(name, self.value.to_string())
-        })
-    }
-
-    /// Converts the repsponse into a list, by splitting the raw value by '/'
-    /// and converting each value into the required type. If any conversion
-    /// fails, an error is returned
-    ///
-    /// # Errors
-    /// If any of the values in the string can not be parsed into the output
-    /// value, the `ParsingError` for that value will be returned
-    pub fn into_list<T: FromStr>(
-        self,
-        name: &'static str,
-    ) -> Result<Vec<T>, SFError> {
-        let x = &self.value;
-        if x.is_empty() {
-            return Ok(Vec::new());
-        }
-        // Trimming ` ` & `\n` is not required. Might remove this later
-        x.trim_matches(|a| ['/', ' ', '\n'].contains(&a))
-            .split('/')
-            .map(|c| {
-                c.trim().parse::<T>().map_err(|_| {
-                    error!(
-                        "Could not convert {name} into list because of {c}: \
-                         {self}"
-                    );
-                    SFError::ParsingError(name, format!("{c:?}"))
-                })
-            })
-            .collect()
-    }
-
-    /// The way keys are parsed will trim some info from the string. The key for
-    /// the player save `ownplayersave` is actually `ownplayersave.playerSave`.
-    /// As this `.playerSave` is not relevant here and not in most cases, I
-    /// decided to trim that off. More common, this is also just `s`, `r`, or a
-    /// size hint like `(10)`. In some cases though, this information can be
-    /// helpful for parsing. Thus, you can access it here
-    #[must_use]
-    pub fn sub_key(&self) -> &str {
-        self.sub_key
-    }
-
-    /// Returns the raw reference to the internal &str, that the server send
-    #[must_use]
-    pub fn as_str(&self) -> &str {
-        self.value
-    }
-}
-
-impl<'a> std::fmt::Display for ResponseVal<'a> {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        f.write_str(self.value)
-    }
-}
-
 #[derive(Debug, Clone)]
 #[allow(clippy::upper_case_acronyms)]
 #[non_exhaustive]
@@ -795,31 +488,6 @@ impl ServerConnection {
     }
 }
 
-fn encrypt_server_request(
-    to_encrypt: String,
-    key: &str,
-) -> Result<String, SFError> {
-    let mut my_key = [0; 16];
-    my_key.copy_from_slice(
-        key.as_bytes()
-            .get(..16)
-            .ok_or(SFError::InvalidRequest("Invalid crypto key"))?,
-    );
-
-    let mut cipher = libaes::Cipher::new_128(&my_key);
-    cipher.set_auto_padding(false);
-
-    // This feels wrong, but the normal padding does not work. No idea what the
-    // default padding strategy is
-    let mut to_encrypt = to_encrypt.into_bytes();
-    while to_encrypt.len() % 16 != 0 {
-        to_encrypt.push(0);
-    }
-    let encrypted = cipher.cbc_encrypt(CRYPTO_IV.as_bytes(), &to_encrypt);
-
-    Ok(base64::engine::general_purpose::URL_SAFE.encode(encrypted))
-}
-
 pub(crate) fn reqwest_client(
     options: &ConnectionOptions,
 ) -> Option<reqwest::Client> {
@@ -833,93 +501,6 @@ pub(crate) fn reqwest_client(
         builder = builder.user_agent(ua);
     }
     builder.default_headers(headers).build().ok()
-}
-
-/// This function is designed for reverseengineering encrypted commands from the
-/// S&F web client. It expects a login resonse, which is the ~3KB string
-/// response you can see in the network tab of your browser, that starts with
-/// `serverversion` after a login. After that, you can take any url the client
-/// sends to the server and have it decoded into the actual string command, that
-/// was sent. Note that this function technically only needs the crypto key, not
-/// the full response, but it is way easier to just copy paste the full
-/// response. The command returned here will be `Command::Custom`
-///
-/// # Errors
-///
-/// If either the url, or the response do not contain the necessary crypto
-/// values, an `InvalidRequest` error will be returned, that mentions the part,
-/// that is missing or malformed. The same goes for the necessary parts of the
-/// decrypted command
-pub fn decrypt_url(
-    encrypted_url: &str,
-    login_resp: Option<&str>,
-) -> Result<Command, SFError> {
-    let crypto_key = if let Some(login_resp) = login_resp {
-        login_resp
-            .split('&')
-            .filter_map(|a| a.split_once(':'))
-            .find(|a| a.0 == "cryptokey")
-            .ok_or(SFError::InvalidRequest("No crypto key in login resp"))?
-            .1
-    } else {
-        DEFAULT_CRYPTO_KEY
-    };
-
-    let encrypted = encrypted_url
-        .split_once("req=")
-        .ok_or(SFError::InvalidRequest("url does not contain request"))?
-        .1
-        .rsplit_once("&rnd=")
-        .ok_or(SFError::InvalidRequest("url does not contain rnd"))?
-        .0;
-
-    let resp = encrypted.get(DEFAULT_CRYPTO_ID.len()..).ok_or(
-        SFError::InvalidRequest("encrypted command does not contain crypto id"),
-    )?;
-    let full_resp = decrypt_server_request(resp, crypto_key)?;
-
-    let (_session_id, command) = full_resp.split_once('|').ok_or(
-        SFError::InvalidRequest("decrypted command has no session id"),
-    )?;
-
-    let (cmd_name, args) = command
-        .split_once(':')
-        .ok_or(SFError::InvalidRequest("decrypted command has no name"))?;
-    let args: Vec<_> = args
-        .trim_end_matches('|')
-        .split('/')
-        .map(std::string::ToString::to_string)
-        .collect();
-
-    Ok(Command::Custom {
-        cmd_name: cmd_name.to_string(),
-        arguments: args,
-    })
-}
-
-fn decrypt_server_request(
-    to_decrypt: &str,
-    key: &str,
-) -> Result<String, SFError> {
-    let text = base64::engine::general_purpose::URL_SAFE
-        .decode(to_decrypt)
-        .map_err(|_| {
-            SFError::InvalidRequest("Value to decode is not base64")
-        })?;
-
-    let mut my_key = [0; 16];
-    my_key.copy_from_slice(
-        key.as_bytes()
-            .get(..16)
-            .ok_or(SFError::InvalidRequest("Key is not 16 bytes long"))?,
-    );
-
-    let mut cipher = libaes::Cipher::new_128(&my_key);
-    cipher.set_auto_padding(false);
-    let decrypted = cipher.cbc_decrypt(CRYPTO_IV.as_bytes(), &text);
-
-    String::from_utf8(decrypted)
-        .map_err(|_| SFError::InvalidRequest("Decrypted value is not UTF8"))
 }
 
 #[derive(Debug, Clone)]
@@ -947,5 +528,131 @@ impl Default for ConnectionOptions {
             expected_server_version: 2005,
             error_on_unsupported_version: false,
         }
+    }
+}
+
+#[cfg(feature = "session")]
+#[derive(Debug)]
+pub struct SimpleSession {
+    session: Session,
+    gamestate: Option<GameState>,
+}
+
+#[cfg(feature = "session")]
+impl SimpleSession {
+    async fn short_sleep() {
+        tokio::time::sleep(Duration::from_millis(fastrand::u64(1000..2000)))
+            .await;
+    }
+
+    /// Creates a new `SimpleSession`, by logging in a normal S&F character
+    ///
+    /// # Errors
+    /// Have a look at `send_command` for a full list of possible errors
+    pub async fn login(
+        username: &str,
+        password: &str,
+        server_url: &str,
+    ) -> Result<Self, SFError> {
+        let connection = ServerConnection::new(server_url)
+            .ok_or(SFError::ConnectionError)?;
+        let mut session = Session::new(username, password, connection);
+        let resp = session.login().await?;
+        let gs = GameState::new(resp)?;
+        Self::short_sleep().await;
+        Ok(Self {
+            session,
+            gamestate: Some(gs),
+        })
+    }
+
+    #[cfg(feature = "sso")]
+    ///  Creates new `SimpleSession`s, by logging in the S&S SSO account and
+    /// returning all the characters associated with the account
+    ///
+    /// # Errors
+    /// Have a look at `send_command` for a full list of possible errors
+    pub async fn login_sf_account(
+        username: &str,
+        password: &str,
+    ) -> Result<Vec<Self>, SFError> {
+        let acc = crate::sso::SFAccount::login(
+            username.to_string(),
+            password.to_string(),
+        )
+        .await?;
+
+        Ok(acc
+            .characters()
+            .await?
+            .into_iter()
+            .flatten()
+            .map(|a| Self {
+                session: a,
+                gamestate: None,
+            })
+            .collect())
+    }
+
+    /// Returns a reference to the game state, if this `SimpleSession` is
+    /// currently logged in
+    #[must_use]
+    pub fn game_state(&self) -> Option<&GameState> {
+        self.gamestate.as_ref()
+    }
+
+    /// Returns a mutable reference to the game state, if this `SimpleSession`
+    /// is currently logged in
+    #[must_use]
+    pub fn game_state_mut(&mut self) -> Option<&mut GameState> {
+        self.gamestate.as_mut()
+    }
+
+    /// Sends the command and updates the gamestate with the response from the
+    /// server. A mutable reference to the gamestate will be returned. If an
+    /// error is encountered, the gamestate is cleared and the error will be
+    /// returned. If you send a command after that, this function will try to
+    /// login this session again, before sending the provided command
+    ///
+    /// # Errors
+    /// - `EmptyResponse`: If the servers response was empty
+    /// - `InvalidRequest`: If your response was invalid to send in some way
+    /// - `ConnectionError`: If the command could not be send, or the response
+    ///   could not successfully be received
+    /// - `ParsingError`: If the response from the server was unexpected in some
+    ///   way
+    /// - `TooShortResponse` Similar to `ParsingError`, but specific to a
+    ///   response being too short, which would normaly trigger a out of bound
+    ///   panic
+    /// - `ServerError`: If the server itself responded with an ingame error
+    ///   like "you do not have enough silver to do that"
+    #[allow(clippy::unwrap_used, clippy::missing_panics_doc)]
+    pub async fn send_command<T: Borrow<Command>>(
+        &mut self,
+        cmd: T,
+    ) -> Result<&mut GameState, SFError> {
+        if self.gamestate.is_none() {
+            let resp = self.session.login().await?;
+            let gs = GameState::new(resp)?;
+            self.gamestate = Some(gs);
+            Self::short_sleep().await;
+        }
+
+        let resp = match self.session.send_command(cmd).await {
+            Ok(resp) => resp,
+            Err(err) => {
+                self.gamestate = None;
+                return Err(err);
+            }
+        };
+
+        if let Some(gs) = &mut self.gamestate {
+            if let Err(e) = gs.update(resp) {
+                self.gamestate = None;
+                return Err(e);
+            }
+        }
+
+        Ok(self.gamestate.as_mut().unwrap())
     }
 }
